picolibc: Allow building picolibc with -Os and -O3
Picolibc can provide both minsize and release builds in the same
toolchain by placing an optional version into a subdirectory
of the default version and allowing applications to provide
--picolibc-buildtype=minsize or --picolibc-buildtype=release to
select the optional version.

Add support in crosstool-ng for this with a new
LIBC_PICOLIBC_ENABLE_TARGET_OPTBOTH flag.

Update the arm-picolibc-eabi sample to use this new flag.

Signed-off-by: Keith Packard <keithp@keithp.com>